checker3/requirements.txt
enochecker3==0.8.1
uvicorn
gunicorn
faker
beautifulsoup4==4.12.3
pillow==10.3.0
pyzbar==0.1.9
segno==1.6.1
